requirements.txt
asyncpraw==7.8.1
discord.py==2.5.0
ffmpeg_python==0.2.0
python-dotenv==1.0.1
yt_dlp==2025.2.19
